Add files via upload
Only the code is included in case to execute the code glf libraries must be included. This project is not entirely complete it can create Bezier curves but functions used to create are not preferred type. In case of how it works when executable code generated there will be white window appear and user must generate dots on screen by using right click.